init operation sorter
